services/ml-processing-service/requirements.txt
fastapi
uvicorn
pydantic
azure-eventhub
azure-storage-blob